feat: add ability to delete submissions

+++ src/app/components/submission-forms/actions.ts
@@ -1,24 +1,18 @@
-'use server';
+"use server";
 
 import { revalidatePath } from "next/cache";
 import postgres from "postgres";
 import { z } from "zod";
 
 let sql = postgres(process.env.PGSQL_HOST!, {
-  ssl: 'allow'
+  ssl: "allow",
 });
-
-interface Submission {
-  submission_id: string
-  submission_name: string
-  submission_datetime: string
-}
 
 export async function createSubmission(
   prevState: {
     message: string;
   },
-  formData: FormData,
+  formData: FormData
 ) {
   const submissionDatetime = new Date().toISOString();
   const schema = z.object({
@@ -53,28 +47,36 @@ export async function createSubmission(
 
 export async function deleteSubmission(
   prevState: {
-    message: string
+    message: string;
   },
   formData: FormData
 ) {
   const schema = z.object({
-    submission_id: z.string().min(1),
+    submission_id: z.coerce.number().gte(1),
     submission_name: z.string().min(1),
   });
-  const data = schema.parse({
-    submission_id: formData.get('submission_id'),
-    submission_name: formData.get('submission_name')
+  const parse = schema.safeParse({
+    submission_id: formData.get("submission_id"),
+    submission_name: formData.get("submission_name"),
   });
+
+  console.log("parse", parse);
+
+  if (!parse.success) {
+    return { message: "Failed to delete submission, parsing error!" };
+  }
+
+  const data = parse.data;
 
   try {
     await sql`
       DELETE FROM submissions
-      WHERE id = ${data.submission_id}
+      WHERE submission_id = ${data.submission_id}
     `;
 
-    revalidatePath('/')
-    return { message: `Deleted submission ${data.submission_name}.`}
-  } catch(e) {
-    return { message: `Failed to delete submission.`}
+    revalidatePath("/");
+    return { message: `Deleted submission ${data.submission_name}.` };
+  } catch (e) {
+    return { message: `Failed to delete submission.` };
   }
 }
